# Deflate speedup3