--- a/src/lz77.ts
+++ b/src/lz77.ts
@@ -8,10 +8,10 @@ const FAST_INDEX_CHECK_MAX = 128;
 const FAST_INDEX_CHECK_MIN = 16;
 const FAST_REPEAT_LENGTH = 8;
 
-function generateLZ77IndexMap(input: Uint8Array) {
-  const end = input.length - REPEAT_LEN_MIN;
+function generateLZ77IndexMap(input: Uint8Array, startIndex: number, targetLength: number) {
+  const end = startIndex + targetLength - REPEAT_LEN_MIN;
   const indexMap: {[key: number]: number[]} = {};
-  for (let i = 0; i <= end; i++) {
+  for (let i = startIndex; i <= end; i++) {
     const indexKey = input[i] << 16 | input[i + 1] << 8 | input[i + 2];
     if (indexMap[indexKey] === undefined) {
       indexMap[indexKey] = [];
@@ -21,9 +21,9 @@ function generateLZ77IndexMap(input: Uint8Array) {
   return indexMap;
 }
 
-export function generateLZ77Codes(input: Uint8Array) {
-  const inputLen = input.length;
-  let nowIndex = 0;
+export function generateLZ77Codes(input: Uint8Array, startIndex: number, targetLength: number) {
+  let nowIndex = startIndex;
+  const endIndex = startIndex + targetLength - REPEAT_LEN_MIN;
   let slideIndexBase = 0;
   let repeatLength = 0;
   let repeatLengthMax = 0;
@@ -34,8 +34,9 @@ export function generateLZ77Codes(input: Uint8Array) {
   const codeTargetValues = [];
   const startIndexMap: {[key: number]: number} = {};
   const endIndexMap: {[key: number]: number} = {};
-  const indexMap = generateLZ77IndexMap(input);
-  while (nowIndex < inputLen) {
+  const indexMap = generateLZ77IndexMap(input, startIndex, targetLength);
+
+  while (nowIndex <= endIndex) {
     const indexKey = input[nowIndex] << 16 | input[nowIndex + 1] << 8 | input[nowIndex + 2];
     const indexes = indexMap[
       indexKey
@@ -66,7 +67,7 @@ export function generateLZ77Codes(input: Uint8Array) {
         || (repeatLengthMax >= FAST_REPEAT_LENGTH && checkCount >= FAST_INDEX_CHECK_MIN)) {
         break;
       }
-      ++checkCount;
+      checkCount++;
       const index = indexes[i];
       for (let j = repeatLengthMax - 1; 0 < j; j--) {
         if (input[index + j] !== input[nowIndex + j]) {
@@ -90,7 +91,8 @@ export function generateLZ77Codes(input: Uint8Array) {
         }
       }
     }
-    if (repeatLengthMax >= 3) {
+
+    if (repeatLengthMax >= 3 && nowIndex + repeatLengthMax <= endIndex) {
       distance = nowIndex - repeatLengthMaxIndex;
       for (let i = 0; i < LENGTH_EXTRA_BIT_BASE.length; i++) {
         if (LENGTH_EXTRA_BIT_BASE[i] > repeatLengthMax) {
@@ -111,5 +113,7 @@ export function generateLZ77Codes(input: Uint8Array) {
       nowIndex++;
     }
   }
+  codeTargetValues.push([input[nowIndex]]);
+  codeTargetValues.push([input[nowIndex + 1]]);
   return codeTargetValues;
 }