Improve deflate speed.
For the measurement, image data of 50 MB is used.

### Restrict loops to speed up https://github.com/zprodev/zlib.es/commit/218afcfe105074a8345bcb312315fb71edc7bc23

|No.|before|after|
|:---:|---:|---:|
|1| Unmeasurable|18488 ms|
|2| Unmeasurable|18605 ms|
|3| Unmeasurable|19456 ms|
|4| Unmeasurable|19019 ms|
|5| Unmeasurable|18501 ms|

### Split compression https://github.com/zprodev/zlib.es/commit/fcbf3d1529f51e83a8fb27c254d036e7e5c1da65

|No.|after|
|:---:|---:|
|1|13447 ms|
|2|13413 ms|
|3|13477 ms|
|4|13569 ms|
|5|13923 ms|